# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          }
        },
        "summary": "Create a new user",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/getRole/{id}": {
      "get": {
        "operationId": "UsersController_findUserRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "b0a4ff3d-9c61-4b59-b1bb-2345f5aa6dcd",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User role fetched successfully"
          }
        },
        "summary": "Get role of a specific user",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "b0a4ff3d-9c61-4b59-b1bb-2345f5aa6dcd",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User fetched successfully"
          }
        },
        "summary": "Get a user by ID",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}/block": {
      "post": {
        "operationId": "UsersController_blockUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "b0a4ff3d-9c61-4b59-b1bb-2345f5aa6dcd",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User blocked successfully"
          }
        },
        "summary": "Block a user",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}/unblock": {
      "post": {
        "operationId": "UsersController_unblockUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "b0a4ff3d-9c61-4b59-b1bb-2345f5aa6dcd",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User unblocked successfully"
          }
        },
        "summary": "Unblock a user",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/tickets/generate-ticket": {
      "post": {
        "operationId": "TicketsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTicketDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ticket generated successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Generate a new ticket",
        "tags": [
          "Tickets"
        ]
      }
    },
    "/tickets/getTicketByCategory/{eventId}": {
      "get": {
        "operationId": "TicketsController_fetchTicketsByCategory",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "248bd4f0-a0e2-4b4c-a503-6ec3b6cedfda",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tickets fetched successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Fetch tickets by event category",
        "tags": [
          "Tickets"
        ]
      }
    },
    "/tickets/scanned": {
      "patch": {
        "operationId": "TicketsController_scan",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanByQrDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ticket scanned successfully"
          }
        },
        "summary": "Scan a ticket by QR code",
        "tags": [
          "Tickets"
        ]
      }
    },
    "/tickets/exportTickets": {
      "get": {
        "operationId": "TicketsController_exportTickets",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Tickets exported successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Export tickets as CSV/Excel",
        "tags": [
          "Tickets"
        ]
      }
    },
    "/tickets/{id}": {
      "delete": {
        "operationId": "TicketsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "123",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket removed successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove a ticket by ID (deprecated)",
        "tags": [
          "Tickets"
        ]
      }
    },
    "/tickets/fetchTickets": {
      "get": {
        "operationId": "TicketsController_fetchTickets",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Tickets fetched successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Fetch all tickets for logged-in user",
        "tags": [
          "Tickets"
        ]
      }
    },
    "/tickets/{eventId}": {
      "get": {
        "operationId": "TicketsController_findEvents",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "248bd4f0-a0e2-4b4c-a503-6ec3b6cedfda",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Events fetched successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Find events by eventId",
        "tags": [
          "Tickets"
        ]
      }
    }
  },
  "info": {
    "title": "Event Management Website",
    "description": "Website for online tickets booking!",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Full name of the user",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "description": "User's email address",
            "example": "john.doe@example.com"
          },
          "password": {
            "type": "string",
            "description": "Password must be at least 8 characters long and contain at least one number",
            "example": "MySecurePass123"
          },
          "roleId": {
            "type": "string",
            "description": "Role ID assigned to the user",
            "example": "dfdeb945-228e-450e-a147-2574f0308967",
            "nullable": true
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "roleId"
        ]
      },
...
      "CreateTicketDto": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "description": "Event ID: ",
            "example": "248bd4f0-a0e2-4b4c-a503-6ec3b6cedfda"
          },
          "pricingId": {
            "type": "string",
            "description": "Pricing Tier Id: ",
            "example": "30931e6f-b36e-4e3b-a10f-8d76024c8d67"
          }
        },
        "required": [
          "eventId",
          "pricingId"
        ]
      },
      "ScanByQrDto": {
        "type": "object",
        "properties": {
          "qrCode": {
            "type": "string",
            "description": "QR code of the ticket",
            "example": "abc123-xyz789"
          }
        },
        "required": [
          "qrCode"
        ]
      }
    }
  }
}